move cassandra to its own module
