+++ buildSrc/src/main/groovy/io.micronaut.build.internal.langchain4j-module-provider.gradle
@@ -4,15 +4,7 @@ plugins {
 
 dependencies {
     annotationProcessor(projects.micronautLangchain4jProcessor)
-    annotationProcessor(mn.micronaut.inject.java)
     annotationProcessor(mnSourcegen.micronaut.sourcegen.generator.java)
-    implementation(platform(libs.langchain4j.bom))
     api(projects.micronautLangchain4jCore)
-    testAnnotationProcessor(mn.micronaut.inject.java)
-    testImplementation(mnTest.micronaut.test.junit5)
-    testRuntimeOnly(mnLogging.logback.classic)
-    testRuntimeOnly(mnTest.junit.jupiter.engine)
     testImplementation(projects.testSuiteChatmodelTck)
-    testImplementation(libs.junit.platform.engine)
 }
-
