+++ settings.gradle.kts
@@ -22,6 +22,7 @@ rootProject.name = "micronaut-langchain4j-parent"
 include("micronaut-langchain4j-bom")
 include("micronaut-langchain4j-processor")
 include("micronaut-langchain4j-core")
+include("micronaut-langchain4j-cassandra")
 if (JavaVersion.current().isCompatibleWith(JavaVersion.VERSION_21)) {
     include("micronaut-langchain4j-jlama")
 }
