+++ src/main/docs/guide/chatModels/chatMemory.adoc
@@ -4,7 +4,7 @@ it relies on you to include the relevant messages in each request for coherent a
 
 Langchain4J provides an API https://docs.langchain4j.dev/tutorials/chat-memory#persistence[`ChatMemory`] to help you manage chat memory. You can provide your own implementation or use one of the provided implementations.
 
-The default implementation of `ChatMemory`, `dev.langchain4j.store.memory.chat.InMemoryChatMemoryStore`, stores `ChatMessage`s in memory.
+The default implementation of `ChatMemory`, `dev.langchain4j.store.memory.chat.InMemoryChatMemoryStore`, stores `ChatMessage` instances in memory.
 
 To use the Redis implementation `dev.langchain4j.community.store.memory.chat.redis.RedisChatMemoryStore`, add the following dependency:
 
@@ -16,7 +16,7 @@ dependency:micronaut-langchain4j-store-neo4j[groupId=io.micronaut.langchain4j]
 
 To use the Cassandra implementation `dev.langchain4j.store.memory.chat.cassandra.CassandraChatMemoryStore`, add the following dependency:
 
-dependency:langchain4j-cassandra[groupId=dev.langchain4j]
+dependency:micronaut-langchain4j-cassandra[groupId=io.micronaut.langchain4j]
 
 The following example shows how to use the `ChatMemoryStore`:
